gdb exp
